requirements.txt
Pillow==10.2.0  # For image processing and EXIF data extraction
pandas==2.2.1   # For CSV handling and data manipulation
tqdm==4.66.2    # For progress bars 